software_team/pyproject.toml
[project]
name = "software-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "duckduckgo-search>=7.5.3",
    "gitpython>=3.1.44",
    "ipykernel>=6.29.5",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.19",
    "langchain-ollama>=0.2.3",
    "langgraph>=0.3.11",
    "llama-index>=0.12.24",
    "llama-index-embeddings-ollama>=0.6.0",
    "llama-index-llms-ollama>=0.5.3",
    "lxml>=5.3.1",
    "markdownify>=1.1.0",
    "matplotlib>=3.10.1",
    "pip>=25.0.1",
    "pydantic-ai-slim[duckduckgo]>=0.0.41",
    "pydantic-ai[logfire]>=0.0.41",
    "pytest>=8.3.5",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["AISoftTeam"]
